// projects/itinerary-lib/src/lib/itinerary-app/_enums/ItenariesRoutes.enum.ts
export enum ItinerariesRoutesEnum {
  DASHBOARD = '/dashboard',
  ITINERARY = '/itineraries',
  ITINERARY_EXPLORE = '/itineraries/explore',
  ITINERARY_FAVOURITES = '/itineraries/favorites',
  ITINERARY_BUILDER = '/itineraries/builder',
  SEARCH = '/search/',
  REQUESTS = '/requests',
  INFO = '/info',
  PROFILE = '/profile',
  ARCHIVES = '/archives/archives',
  COMMENTS = '/itineraries/comments',
  EXPLORE = 'explore',
  BUILDER = 'builder',
  CREATE_ITINERARY = 'create',
  CREATE_ITINERARY_ORIGIN = 'origin',
  CREATE_ITINERARY_DESTINATION = 'destination',
  SELECT_ITINERARY_EVENTS = 'select-events',
  SELECT_ITINERARY_ADD_PLACE = 'add-place',
  SELECT_ITINERARY_PLACE_BASIC_INFO = 'info',
  SELECT_ITINERARY_PLACE_DATETIME = 'datetime',
  SELECT_ITINERARY_PLACE_MEMBERS = 'members',
  CREATE_ITINERARY_USER_SEARCH = 'user-search',
  CREATE_ITINERARY_MEMBER_DETAILS = 'member-detail',
  CREATE_ITINERARY_DRAW_MODE = 'draw',
  CREATE_ITINERARY_DRAW_MODE_CREATE = 'create',
  CREATE_ITINERARY_DRAW_MODE_PREVIEW = 'preview',
  CREATE_ITINERARY_NO_VISUAL_FILTER_MODE = 'no-filters',
  CREATE_ITINERARY_NO_VISUAL_FILTER_MODE_CREATE = 'create',
  CREATE_ITINERARY_NO_VISUAL_FILTER_MODE_PREVIEW = 'preview',
  CREATED_ITINERARY = 'created-itinerary',
  FAVORITES = 'favorites',
  STORIES = 'stories',
  COMMENTS_ROUTE = 'comments',
  NOTIFICATIONS = 'notifications',
  COMMENT_UPDATE_REQUEST = 'comment-update-request',
  ITINERARY_MEMBERS = 'members',
  ALL_INVITES_SENT = 'invites',
  SEND_INVITE = 'invite',
  All_PEOPLE = 'people',
  ITINERARY_FRIENDS = 'friends',
  LIST_FRIENDS = 'list-friends',
  ITINERARY_DETAIL = 'detail',
  ITINERARY_USERS = 'users',
  ITINERARY_UPVOTE_USERS = 'upvote-users',
  ITINERARY_WATCHING_USERS = 'watching-users',
  ITINERARY_AUTHOR_HISTORY = 'author-history',
  ITINERARY_SUMMARY = 'summary',
  ITINERARY_MAP = 'map',
  ITINERARY_MORE_DETAIL = 'more',
  BRANCHED_ITINERARIES = 'branch',
  COMMENT_SEARCH = 'comment-search',
  FRIEND_REQUESTS = 'friend-requests',
  SEARCH_FRIEND = 'search-friend',
  COMMENT_REPLIES = 'comment-replies',
  ADVANCED_FILTERS = 'advanced-filters',
  SEND_FRIEND_REQUEST = 'send-friend-request',
  LOCATION_PICKER = 'location-picker',
  SUGGESTED_EVENTS_LOCATION = 'suggested-events-location',
  SUGGESTED_EVENTS_ROUTE_TYPE = 'suggested-events-route-type',
  SUGGESTED_EVENTS_ROUTE_RADIUS = 'suggested-events-route-radius',
  CREATED_ITINERARY_PREVIEW = 'created-itinerary-preview',
  CREATED_ITINERARY_DETAIL = 'detail',
  CREATED_ITINERARY_MAP = 'map',
}